os: relocate renv / rmodule / rtty under subdir
os/ already carries rproc.h, rsignal.h and rsys.h - the OS-system
facing primitives. Environment variables, dynamic loading
(dlopen / LoadLibrary) and TTY detection fit the same shape and
weren't at the top level for any architectural reason.

Affects 18 #include sites plus the meson source list. The .c
files move alongside their headers (rlib/{renv,rmodule,rtty}.c
-> rlib/os/).
