+++ rlib/meson.build
@@ -8,7 +8,7 @@ rlib_sources = [
   'rclock.c',
   'rcpufeatures.c',
   'rcrc.c',
-  'renv.c',
+  'os/renv.c',
   'rio.c',
   'net/riosocket.c',
   'rjson.c',
@@ -20,7 +20,7 @@ rlib_sources = [
   'rmem.c',
   'rmemfile.c',
   'rmemscan.c',
-  'rmodule.c',
+  'os/rmodule.c',
   'crypto/rmsgdigest.c',
   'rpoll.c',
   'rprng-kiss.c',
@@ -36,7 +36,7 @@ rlib_sources = [
   'rthreadpool.c',
   'rthreads.c',
   'rtime.c',
-  'rtty.c',
+  'os/rtty.c',
   'ruri.c',
   'asn1/rasn1bin.c',
   'asn1/rasn1binenc.c',
