+++ rlib/rlog.c
@@ -24,12 +24,12 @@
 #include <rlib/file/rfile.h>
 #include <rlib/os/rproc.h>
 
-#include <rlib/renv.h>
+#include <rlib/os/renv.h>
 #include <rlib/rmem.h>
 #include <rlib/rstr.h>
 #include <rlib/rthreads.h>
 #include <rlib/rtime.h>
-#include <rlib/rtty.h>
+#include <rlib/os/rtty.h>
 
 
 rauint _r_log_level_min = R_LOG_LEVEL_DEFAULT;
